fix(consensus): correct Option type in Header::size() method
Fix type mismatch in Header::size() where Option<u128> was used
for base_fee_per_gas, blob_gas_used, and excess_blob_gas fields
that are actually Option<u64> in the struct definition. This was
overestimating the memory size by 48 bytes.

+++ crates/consensus/src/block/header.rs
@@ -254,9 +254,9 @@ impl Header {
         mem::size_of::<u64>() + // timestamp
         mem::size_of::<B256>() + // mix hash
         mem::size_of::<u64>() + // nonce
-        mem::size_of::<Option<u128>>() + // base fee per gas
-        mem::size_of::<Option<u128>>() + // blob gas used
-        mem::size_of::<Option<u128>>() + // excess blob gas
+        mem::size_of::<Option<u64>>() + // base fee per gas
+        mem::size_of::<Option<u64>>() + // blob gas used
+        mem::size_of::<Option<u64>>() + // excess blob gas
         mem::size_of::<Option<B256>>() + // parent beacon block root
         mem::size_of::<Option<B256>>() + // requests root
         self.extra_data.len() // extra data
